requirements.txt
minio
celery
psycopg2-binary
SQLAlchemy
importlib_metadata